Fixed the cmake shared library linking and memory bugs.
Also debugged heavily in valgrind and it seems like the dubious exit bug
is finally gone now. Feeling good about this!

Also fixed some errors with updating the easylogging++ library.

The only problem left are some memory leaks in the SDLEventHandler
(according to Valgrind) - the getEvent() function is not memory leak
free. I don't see a possible bug in there now, but that can change in
the future.